feat(nvim): Initial configuration
This handles the setup of our neovim configuration.

We've gone for a big bang approach, as we've needed to play around with
our settings and make sure that we've at least got a close approximation
of the setup we originally had within vim.

We also spent a few days fleshing out what our PDE should feel like and
what we _really_ need, as opposed to simply bringing in all of the vim
plug-ins we previously had.

There's still some improvements we could make, but waiting until this
setup is perfect will also introduce errors where they're not needed.

Signed-off-by: Yomi Colledge <baphled@boodah.net>