feat: Initial commit
Here we make sure that we have all of the basics to get started with our
new setup.

This includes all the plug-ins that we know we need, along with some
basic configuration settings for plug-in that we have been testing over
the last couple of weeks.

We've omitted configuring all of the plug-ins, for now, as we will be
introducing them incrementally as we gain more comfortability using
them.

Signed-off-by: Yomi Colledge <baphled@boodah.net>