feat(coverage): Automatically enable
Setup `nvim-coverage` to intelligently look up the file in a buffer and
check its mime type against the languages types nvim-coverage supports.

Once this is done we then check that we have a code coverage JSON file.

If this is the case we then load the code coverage file and
automatically show it.

This allows us to always have code coverage within our PDE without
having to think or worry about whether it has been enabled.

As we don't usually write code without tests, this means that when ever
we're making changes to a code base we always have in-line coverage
enabled helping us to focus on getting stuff down and not swapping
between the terminal or a browser.

We also add `<leader>cs` which allows us to get a summary of our code
coverage.

Side note: It should now be possible to display our coverage within our
greeter. Perhaps this is something we can explore at a later date.

Signed-off-by: Yomi Colledge <baphled@boodah.net>