feat(neotest): Support running Ruby and JS tests
Even though it's possible to run our tests outside of our editor it's
more practical to have this functionality within our PDE.

For now, we'll only focus on getting Ruby and Jest running within our
PDE, but it is possible to extend this to other languages.

We're using `neotest-jest` and `neotest-ruby` to do most of the hard
work and we've also add a handful of mapping to allow us to run our
tests in a similar way to how we had them running within Vim.

Commands:
  * <leader>tw  - Watch tests (Jest Only)
  * <leader>a   - Run all tests
  * <leader>t   - Run tests within file
  * <leader>T   - Run all tests
  * <leader>tx  - Cancel tests
  * <leader>ts  - Test Summary
  * <C-T>       - Test output

We can later extend this to cover the other languages we commonly use,
but as we tend to focus on these two languages. This should do for now.

Signed-off-by: Yomi Colledge <baphled@boodah.net>