feat(nvim-dap): Add DAP support for JS and Ruby
As Ruby is the most commonly used language, we introduce `ruby-dap` so
that we can debug our Ruby code.

Similarly to LSP, DAP is a common layer, which variously languages can
tap into. So, it is very possible for us to integrate support for other
languages at a later time.

Signed-off-by: Yomi Colledge <baphled@boodah.net>