feat(LSP): Integrate LSP into our PDE
This allows us to leverage the Ruby LSP,along with a handful of commonly
used languages.

Again, as Ruby is the most commonly used language we use it makes sense
to make sure that we have this functionality from the outset. We also
add clang, as we can leverage this for for our C++ code bases.

* "K" - Symbol Hover
* "<leader>gws" - Workspace symbol
* "<leader>gd" - Diagnostic
* "[d" - Previous diagnostic
* "]d" - Next diagnostic
* "<leader>gca" - code rename
* "<leader>grn" - code action new name
* "<C-H>" - Signature help
* "gr" - reference in Trouble
* "gd" - definition in Trouble
* "<leader>xw" - Open workspace diagnostics in Trouble
* "<leader>xd" - Open document diagnostics in Trouble

Signed-off-by: Yomi Colledge <baphled@boodah.net>