cix(copilot): Ensure `nvim-cmp` doesn't overwrite copilot maps
Our initial setup was done in a way were copilot was unable to be
triggered, as nvim-cmp takes control of <Tab> and <S-Tab>. It seems that
the accepted way of resolving this is to take the mappings from LSP, map
them to nvim-cmp and then disable these keys.

Here, we make sure that we take this approach and tweak our mappings.
Whilst we are at it we also add a border to nvim-cmp so that we make our
mapping pop a little bit more.

Signed-off-by: Yomi Colledge <baphled@boodah.net>