refactor(remap): Extract remaps into logical chunks
As our remaps grow it's becoming harder to keep track of where they
are. Extracting them into logical chunks allows us to be able to jump
straight to them quicker and makes it more intuitive to add and remove
them by finding the associated remap file.

Here we reorganise our remaps so that we have a basic file structure for
these mappings and remove the old file with this new structure.

Signed-off-by: Yomi Colledge <baphled@boodah.net>