feat(which-key): Integrate which-key
This plug-in allows us to display a pop-up which displays all of our key
bindings. As we have a number of customised key bindings, it will prove
helpful to integrate this functionality into our workflow.

This plug-in already taps into a number of our existing plug ins, further
enhancing our experience. One of these being our spellchecker
suggestions.

This along makes it a lot more fluid to make quick changes
with improving the way we work within our PDE.

As this is the case, we also refactor our telescope mappings so that
they are handled by which-key. This opens us up to a whole range of key
bindings that we never considered, along with the added benefit of
providing us with feedback as we type our new mappings.

We've extracted each of our telescope mapping to work with new mnemonics.
This shouldn't hinder us much as we have the assistance of which-key to
remind us of what keys we need to press

Signed-off-by: Yomi Colledge <baphled@boodah.net>