feat(notify) Integrate nvim-notify
The main aim for introducing this into our PDE is to encourage us to get
feedback on things like testing and debugging.

Currently, we've only introduced this into our a couple of debugging
keymaps and will gradually expand it into other aspects of our workflow.

The idea being that it can help to notify us of aspects of our workflow
where we are waiting or want confirmation that something has changed.

We don't want to go crazy with this, as notifications can easily become
distraction, but if it's done correctly, this can really help us to get
instant feedback on areas of our workflow where it is important to do so.

Signed-off-by: Yomi Colledge <baphled@boodah.net>