feat(lualine): Add winbar
As we have a global status bar, knowing which file we're working with
can get confusing.

To resolve this, we can simply add a winbar status. This functionality
comes with lualine, and allows us to define a bar for each of our vim
windows. Leading to us having each window clearly labelled.

Signed-off-by: Yomi Colledge <baphled@boodah.net>