feat(lsp): Auto-format on save
We want to be able to not think about formatting any more. Often, we
find ourselves messing around with it whilst trying to get thoughts out
of our head.

Now that we have LSP, and have played around with having a key binding
for handling this task. We think it makes much more sense to trigger
  formatting anytime a file has been saved.

Ideally, we would have this triggered when ever we come out of insert
mode, but for now it is enough to just trigger it whenever we save.

This way, we can focus on getting our thoughts out without having to be
interrupted by ill formatted files.

Signed-off-by: Yomi Colledge <baphled@boodah.net>