feat(packer): Source file on save
Often times we want to source Packer before we call `PackerSync`.
There's been a number of times where we've forgotten that this is the
case, and ended up closing neovim. We're now pretty good at remembering
to source the file, but it is still nicer not to have to think about
this.

Here we create an auto command that checks whether the buffer is
`packer.lua` and if it is to source the file when wee have saved it.

We could extend this a little further, so that we always run
`PackerSync` but for now this will be enough of a change.

Signed-off-by: Yomi Colledge <baphled@boodah.net>