refactor(which-key): Add notify to neotest commands
Sometimes these commands can take a moment to get started. So, to help
provide feedback that they are actually running we add notify actions to
each command.

This can help let us know when something has started, so we don't end up
calling the same command multiple times.

It would be nice if we could also have a notify event for for when our
tests pass, with some basic information, but that is out of scope of
this change.

Signed-off-by: Yomi Colledge <baphled@boodah.net>