feat(neotest): Improve how we handle JS based tests
Until now, we've assumed that we're always running our Jest based tests
via vue-cli-service. This isn't always the case, so we've updated our
`jestCommand` property to be a bit smarter.

This will likely need tweaking in the future, but for now we've made it
smart enough to look for `vue-cli-service` or `jest` within
`node_modules/bin` and use the one that's appropriate.

As VueJS based projects typically use both, it's important to check for
`vue-cli-service` first, and then fall back on `jest`. So we have made
sure this is the case when doing our checks.

Signed-off-by: Yomi Colledge <baphled@boodah.net>