feat(lsp-notifty): Integrate nvim-notify with lsp-zero
This required a bit of tweaking to allow for this functionality to work.
We've confirmed that it's working with `luals`, although it does not
seem to be triggered with Ruby project. This may also apply to other
LSPs we've got setup.

For now, it's good enough to have the functionality there, and then we
can extend upon it later, when we have enough space to do so.

Signed-off-by: Yomi Colledge <baphled@boodah.net>