fix(packer): Auto saving an syncing was breaking our plug-ins
We're not entirely sure why this is the case, but once we fixed the typo
we noticed that all of our packages would be removed.

It's likely that this is due to the fact that the autocommand was setup
before registering any of the packages. Although this wasn't confirmed.

For now, we'll remove this functionality, so that we don't cause further
issues.

Signed-off-by: Yomi Colledge <baphled@boodah.net>