feat(neogit): Add plug-in
This could serve as a replacement for `fugitive`. Although `fugitive`
has served us extremely well. We want a git plug-in that can tap into
the `vim.notify` layer and provide notifications of git based changes.

We will need to experiment with it for a while, especially as we've
been using fugitive for so long, but we'll introduce it here to get a
feel for the pros and cons.

Signed-off-by: Yomi Colledge <baphled@boodah.net>