Change how we handle automatically formatting a file.
Attempting to do this with an autocommand caused strange side effects
where if we had the same buffer open twice then one would jump to the
beginning of the file. This is confusing and leads to the potential of
losing changes if we're bouncing between the two buffers.

lsp-zero had a command for handling the ability to auto-format files, so
we swap out our version for the cleaner approch provided by lsp-zero.

Signed-off-by: Yomi Colledge <baphled@boodah.net>