feat(which): Add <C-b> for closing all buffers
Sometimes we want to create a clean slate, and get rid of all buffers
that we currently have available. It's true that we could do this
manually, but it is much easier to add a key binding which allows us to
do this easily.

Here we add <C-b> to allow us to do _just_ this and ensure that it is
something that's easy to accomplish.

Signed-off-by: Yomi Colledge <baphled@boodah.net>