fix(nvim-notify): Get rid of `lsp-notify`
We seem to be having issues with `lsp-notify` where it constantly errors
when we prematurely close a buffer. We've not managed to get to the
bottom of why this is happening, and the feature is _purely_ cosmetic.
So, we've decided to get rid of it for now.

It would be nice get bring it back at some point, but we'll need to
confirm that a) the issue is completely resolved, or b) we've had time
to investigate the issue and resolve it.

Signed-off-by: Yomi Colledge <baphled@boodah.net>