fix(which-key): Tweak our LSP keybindings
Improve our keybindings so that we have a more intuitive experience.

When we're doing a signature look up we want to be able to do this in
insert mode, and we also want to be able to do so easily. For this we've
created <Ctrl-h>.

We've also rearranged the order of some of our keys so that we maintain
an alpha-numeric order to them.

Signed-off-by: Yomi Colledge <baphled@boodah.net>