refactor(which-key): Move map definitions to separate files
As we now have quite a lot of custom key maps it is becoming cumbersome
to keep track of what changes map to what plug-in.

We also want to be able to easily move these around, if/when we decide
to swap `packer` for `lazy`.

Doing this in one big change would be a pain, so it makes sense to take
this initial step to make it easier to migrate these mappings later on.

Signed-off-by: Yomi Colledge <baphled@boodah.net>