feat(noice): Add package
To enhance our UI we've introduced `noice` to help tie together our
command interactions with `nvim-notify`.

There is an existing issue where we're not able to sign GPG commits, but
this is a known issue. We're hopeful that this will be resolved with a
fix from NeoVim.

With this aside, we're now able to customise our command based
interactions within NeoVim and have their updates sent directly to
`nvim-notify`.

Signed-off-by: Yomi Colledge <baphled@boodah.net>