chore(other): Define package lock files for JavaScript and Ruby
We often want to jump between `package.json` and `package-lock.json` or
`Gemfile` or `Gemfile.lock`.

If we wanted to do this, previously, we'd need to either use a file
navigator or `Telescope`. This can be cumbersome and takes more time
than what we originally had setup in our previous vim settings.

Here, we add this ability through `other.vim`, which already gives us
the ability to define alternative files or us. This means that we can
use our existing key bindings and not have to think about the file type
and rely upon a convention we're already very used to.

Signed-off-by: Yomi Colledge <baphled@boodah.net>