feat(harpoon): Add harpoon
We want a cleaner way of navigating between a small collection of files.

We're relying _far_ too much on what's stored in our buffers and we tend
to fall back to either Telescope or NVimTree.

To further improve our productivity, we've introduced `harpoon` and
setup some custom key bindings to help us to organise a small collection
of files and be able to navigate them.

As we're extensively using `which-key`, we've decided to integrate these
bindings into our new `which-key` collection.

This means that when we decide to migrate to `lazy` it should be a lot
easier to move our key bindings into our plug-in setup files, and
_finally_ have everything associated to a plug-in living in one place.

Signed-off-by: Yomi Colledge <baphled@boodah.net>