feat(nvim-dap): Add basic debugging for javascript and typescript
This allows us to debug our code in a similar way to what we're already
able to do with Ruby.

There is a caveat here. We've not managed to get this to work as
seamlessly as the experience we have with Ruby, as `nvim-dap` hangs, and
doesn't terminate, when we're finished stepping through the code.

Signed-off-by: Yomi Colledge <baphled@boodah.net>