feat: Replace packer with lazy
We've had a few issues whether we go to update our plug-ins and we end up
with a partially broken system.

We've done some digging an discovered that lazy provides SHAs of the last
commit, which will help us to have an easy wall to make sure were
working with package versions that are stable. So here we replace packer
with lazy and make all the necessary changes to completely remove
packer.

Signed-off-by: Yomi Colledge <baphled@boodah.net>