chore(lsp-zero): Update with Lazy in mind
Now that we've introduced Lazy, we've not some _slight_ issues that need
to be resolved. Firstly, we need to migrate from using `lsp-zero` for
handling languages to `Mason`. Next we had some issues with `lua_ls`
throwing some errors.

Here, we make sure that our config is up to date.

Signed-off-by: Yomi Colledge <baphled@boodah.net>