chore(lazy): Update version of lazy
We played with removing _all_ of our dependencies, to ensure that when
we're working with new instances, that we're able to set up those
instances without any issues. This meant that Lazy bumped it's version.

This isn't a big deal, but we want to make sure we capture this, before
attempting to upgrade any of our other dependencies.

Signed-off-by: Yomi Colledge <baphled@boodah.net>