feat(package): Add Commentary
We used to have this, in our previous setup, but we ended up removing
it. This then, we've missed having it in our setup and could _really_
benefit from bring it back in.

Here, we re-introduce it and bind it to the _old_ binding we used to
have, so that we don't have to re-learn a new binding.

Signed-off-by: Yomi Colledge <baphled@boodah.net>