feat(Lazy): Add image.nvim
This allows us to be able to view images within our markdown documents.

The _really_ cool thing with this is that not only can we view our
images but we can get a feel for what the images look like without
having to view a GUI.

We're currently leveraging kitty here, which is fine as that's our
primary terminal, but we're also able to use ueberzugpp if we later
decide that this isn't working out.

Signed-off-by: Yomi Colledge <baphled@boodah.net>