feat(markdown): Update packages to help us to write markdown based content
We've been using Obsidian to manage our notes for a few days now and we
like _most_ of what it provides. One of the things we've spent a lot of
time on is formatting tables, which can be a _serious_ pain to deal
with. It turns out that `mkdnflow` actually provides functionality to do
this better than tabular (although, not perfect) so we've decided to
re-introduce it.

We also need to be able to add emojis to our content, especially as
we've found this beneficial in the past, so we have introduced
`cmp_emoji`.

Another areaa that has been lacking is having our markdown pages
automatically wrap our documentation which especially breaks our
tables. So we need to change these setting to make sure that we don't
accidentally break our formatting.

Finally, we have introduced zen-mode. We need to be aware that zen-mode
can break image.nvim, but for the most part we see it to be a benefit.
Especially when we're focusing on a piece of documentation and want to
have no distractions.

Signed-off-by: Yomi Colledge <baphled@boodah.net>