feat(obsidian): Add basic key bindings
We want to be able to have access to some of the core functionality that
Obsidian provides _anywhere_ within our editing experience. This will
help us to be able to quickly make notes whilst we're working on various
types of things, whilst being able to easily jump back when needed.

Signed-off-by: Yomi Colledge <baphled@boodah.net>