fix(image.nvim): Only render images is at cursor
We love this plug-in but it can cause issues with things link zen-mode.
Having images only display when our cursor is on the image link means
that not only is it disabled in zen-mode but it also means we're able to
focus on writing, whilst still being able to view images when we want
to.

The only caveat is that we need to make sure our cursor is not on the
image whilst in zen-mode otherwise the plug-in messes with our layout.

Signed-off-by: Yomi Colledge <baphled@boodah.net>