fix(lualine): Make sure our globalstatus works as expected
Our previous config caused issues with globalstatus not being respected.
This meant that all of our windows displayed a status line, instead of
having one status line that updated depending on the window we were
working on.

This change correct this, so that our status line works as we intended
it to.

Signed-off-by: Yomi Colledge <baphled@boodah.net>