feat(trouble): Configure so that we're able to view symbols and diagnostic
We want to improve the layout of our PDE so that we're able to navigate
between code better. We're able to do this with the help of trouble and
a bit of customisation.

We've also tweaked edgy to allow us to fix the symbols and diagnostic
windows to the right. Whilst the other trouble windows are displayed at
the bottom.

We've also tweaked neotest, so that when our tests have finished running
qflist pops up (using trouble). So that we can easily jump between our
failed tests.

Signed-off-by: Yomi Colledge <baphled@boodah.net>