fix(Other): Improve key bindings
We've been using the same key bindings for over a decade, but _now_ that
we're comfortable with nvim; it feels cumbersome to continuously use
these _clunky_ commands.

We've modified our key bindings, so that we can still use the legacy
version but also added the ability to use <M-a> as an alternative. We'll
play round with these for a while to see if this feels more natural.

Signed-off-by: Yomi Colledge <baphled@boodah.net>